requirements.txt
rclpy
fastapi
uvicorn[standard]
pydantic
python-dotenv
psycopg2-binary
sqlalchemy
pytest
pytest-cov
pyyaml
